skiboot: resplit secvar.h
In skiboot, secvar.h is in both libstb/secvar and in include/.
In secvarctl we merged them.
Unmerge them now in preparation for bringing in all of skiboot.
