+++ Makefile
@@ -2,7 +2,8 @@
 # Copyright 2021 IBM Corp.
 #_*_MakeFile_*_
 CC = gcc 
-_CFLAGS = -MMD -O2 -std=gnu99 -I./ -Iinclude/ -Iexternal/skiboot/ -Wall -Werror
+_CFLAGS = -MMD -O2 -std=gnu99 -I./ -Iinclude/ -Iexternal/skiboot/ \
+	  -Iexternal/skiboot/include -Wall -Werror
 
 DEBUG ?= 0
 ifeq ($(DEBUG),1)
