+++ backends/edk2-compat/edk2-svc-read.c
@@ -10,7 +10,7 @@
 #include <stdlib.h>
 #include <argp.h>
 #include "crypto/crypto.h"
-#include "external/skiboot/include/secvar.h" // for secvar struct
+#include "external/skiboot/libstb/secvar/secvar.h" // for secvar struct
 #include "backends/edk2-compat/include/edk2-svc.h" // include last, pragma pack(1) issue
 
 static int readFiles(const char *var, const char *file, int hrFlag,
