+++ backends/edk2-compat/edk2-svc-verify.c
@@ -7,7 +7,7 @@
 #include <unistd.h> // has read/open functions
 #include <argp.h>
 #include "external/skiboot/include/opal-api.h"
-#include "external/skiboot/include/secvar.h"
+#include "external/skiboot/libstb/secvar/secvar.h"
 #include "backends/edk2-compat/include/edk2-svc.h"
 
 struct Arguments {
