+++ external/skiboot/libstb/secvar/backend/edk2-compat-process.h
@@ -18,15 +18,14 @@
 #include <device.h>
 #include "libstb/crypto/pkcs7/pkcs7.h"
 #include "external/skiboot/libstb/secvar/backend/edk2.h"
-#include "opal-api.h"
+#include "opal-api.h"*/
 #include "../secvar.h"
-#include "../secvar_devtree.h"*/
+/*#include "../secvar_devtree.h"*/
 
 //added by nick child/dja
 #include <ccan/short_types/short_types.h>
 #include "external/skiboot/include/opal-api.h"
 //#include "external/extraMbedtls/include/pkcs7.h"
-#include "external/skiboot/include/secvar.h"
 #include "external/skiboot/libstb/secvar/backend/edk2.h"
 
 #define __unused		__attribute__((unused)) //ADDED BY NICK CHILD
